+++ R/Table.R
@@ -161,11 +161,11 @@ createFields <- function(con, fields, field.types, row.names) {
   if (!is.null(field.types)) {
     is_field <- names(field.types) %in% names(fields)
     if (!all(is_field)) {
-      stop(
-        sprintf("Columns in `field.types` must be in the input, missing columns:\n%s",
+      warning(
+        sprintf("Some columns in `field.types` not in the input, missing columns:\n%s",
           paste0("  - '", names(field.types)[!is_field], "'", collapse = "\n")
         ),
-      call. = FALSE)
+      call. = FALSE, immediate. = TRUE)
     }
 
     fields[names(field.types)] <- field.types